Remove client from docker-dev, split out docker-local
Client isn't useful in docker dev, as we recommend you start the client
manually on your local machine for hot reloading etc... So removed it
from docker-dev so it doesn't waste time building something we don't
use.

Also created a docker-local which does launch the client, but only use
this for if you're intending on running it to use locally.